requirements.txt
pandas
numpy
requests
pyyaml